+++ src/index.css
@@ -24,6 +24,11 @@ input:focus {
   margin: 0 auto;
 }
 
+.done {
+  text-decoration: line-through;
+  opacity: 0.5;
+}
+
 .li {
   list-style: none;
   padding: 10px;
