+++ src/pages/Todo/TodoListPage.tsx
@@ -1,13 +1,7 @@
-import { Title, Spacing, Button } from 'components';
-import { useState } from 'react';
+import { Title, Spacing, Button } from 'components/UI';
+import TodoItem from 'components/Todo/TodoItem';
 
 export default function TodoListPage() {
-  const [isUpdate, setIsUpdate] = useState(false);
-  const handleUpdate = () => {
-    setIsUpdate(prev => !prev);
-  };
-  console.log(isUpdate);
-
   return (
     <>
       <Title>todos</Title>
@@ -20,20 +14,10 @@ export default function TodoListPage() {
       </form>
       <Spacing size={17} />
       <ul className="ul">
-        <li className="li">
-          <label>
-            <input type="checkbox" className="checkbox" />
-            <span>TODO 1</span>
-          </label>
-          <div>
-            <Button data-testid="modify-button" type="light" onClick={handleUpdate}>
-              수정
-            </Button>
-            <Button data-testid="delete-button" type="danger">
-              삭제
-            </Button>
-          </div>
-        </li>
+        <TodoItem />
+        <TodoItem />
+        <TodoItem />
+        <TodoItem />
       </ul>
     </>
   );
